refactor: migrate ButtonGroup test to testing-library
